src/main/java/Skypro/Coursework2/service/ExaminerServiceImpl.java
package Skypro.Coursework2.service;

import Skypro.Coursework2.dto.Question;
import Skypro.Coursework2.exception.NotEnoughQuestions;
import org.springframework.stereotype.Service;

import java.util.Collection;
import java.util.HashSet;
import java.util.Set;

@Service
public class ExaminerServiceImpl implements ExaminerService {
    private final QuestionService questionService;
    ExaminerServiceImpl(QuestionService questionService) {
        this.questionService = questionService;
    }
    @Override
    public Collection<Question> getQuestions(int amount) {
        Set<Question> questionSet = new HashSet<>();
        if (questionService.getAll().size() < amount) {
            throw new NotEnoughQuestions("Not questions is collection");
        }
        while (questionSet.size() < amount) {
            questionSet.add(questionService.getRandomQuestion());
        }
        return questionSet;
    }
}
